Adding `compilerIO` flag to set subproc STDIO
Currently the option for STDIO is coded to be `inherit`, which is a good
default for most use cases.  However, sometimes a consumer would like to
capture the STDIO for processing or analysis, so adding this flag allows
them to pass `pipe` to capture, or `ignore` to discard the STDIO
altogether.